# Slow performance of `NEVec` and `NonEmptyIterator`