--- a/src/lib.rs
+++ b/src/lib.rs
@@ -112,6 +112,7 @@ pub mod map;
 pub mod set;
 pub mod slice;
 pub mod vector;
+pub mod vec2;
 
 #[cfg(feature = "indexmap")]
 pub use index_map::NEIndexMap;