@fosskers 2024-04-08
In general I am open to redesigns that improve performance / usability as you've demonstrated here. 

I'm going on an extended trip next week however, and will be gone until the end of May. I'd like to revisit this in detail then, since I can't devote the time to it this week. How does that sound?