@rinde 2024-04-08
I understand. That is unfortunate (for me) since I hoped to getting this fixed soon, but nice for you to go on a long trip 👍 . I will probably continue working in a separate PR and maybe create a temporary fork so that we can use it internally. Are you still planning a release before you leave that includes the [arrays PR](https://github.com/fosskers/nonempty-collections/pull/20)?

Have a nice trip!