@fosskers 2024-04-09
I've just released `0.2.5` with the array improvements.

> I will probably continue working in a separate PR and maybe create a temporary fork so that we can use it internally.

Please feel free, although let's definitely continue collaboration once I return. I am keen for a better design, and these next few weeks could be a good opportunity for you to test the new design internally.