@rinde 2024-04-09
Thank you for the release! 

> Please feel free, although let's definitely continue collaboration once I return. I am keen for a better design, and these next few weeks could be a good opportunity for you to test the new design internally.

Yes that is indeed my intent. I will create an improved version and test it internally. When you're back we can discuss my findings and see how we can best continue our collaboration to further improve the library.